+++ src/components/floorplans/SortableItem.jsx
